feat: Implement budget and badge data models
This commit introduces the data models for budgets and badges and integrates them into the `UserItem` and `DashboardActivity`.

This includes:
- **`UserItem.java`**:
  - Renamed `legalName` to `email` for clarity.
  - Added a `Budget` object and a list of `Badges` to each user.
  - New users are now initialized with a default budget and a set of badges.

- **`Budget.java`**:
  - New data model to manage a user's budget.
  - It contains a list of `BudgetLimitCategory` and `Transaction` objects.
  - Provides methods to calculate total budget, total spent, and spending on "wants".

- **`BudgetLimitCategory.java`**:
  - Changed `double` types for `limitAmount` and `amountSpent` to `int`.

- **`Badges.java`**:
  - Added `badgeName` and `badgeIconId` fields.
  - The `addxp` method is renamed to `addXp`.

- **`DashboardActivity.java`**:
  - Now receives `UserItem` data from an Intent.
  - Dynamically populates progress bars and text views with the user's budget data, showing overall spending and the top two spending categories.

- **New Assets**:
  - Added `budget_warrior.png` and `literacy_icon.png` (renamed from `literacy_scholar.png`) as drawable resources for badges.

+++ app/src/main/java/com/example/hackathon2025/models/BudgetLimitCategory.java
@@ -4,11 +4,11 @@ import java.io.Serializable;
 
 public class BudgetLimitCategory implements Serializable {
     private final String categoryName;
-    private double limitAmount;
-    private double amountSpent;
+    private int limitAmount;
+    private int amountSpent;
     private boolean isWant;
 
-    public BudgetLimitCategory(String name, double limit, double spent, boolean isWant) {
+    public BudgetLimitCategory(String name, int limit, int spent, boolean isWant) {
         this.categoryName = name;
         this.limitAmount = limit;
         this.amountSpent = spent;
@@ -30,11 +30,11 @@ public class BudgetLimitCategory implements Serializable {
         return isWant;
     }
 
-    public void setLimitAmount(double limitAmount) {
+    public void setLimitAmount(int limitAmount) {
         this.limitAmount = limitAmount;
     }
 
-    public void addToAmountSpent(double amountSpent) {
+    public void addToAmountSpent(int amountSpent) {
         this.amountSpent += amountSpent;
     }
 }
